refactor: Keep strings aligned in separate columns
While keeping the string data in a CSV-style table is fine during
parsing, it required the usage of lists to make ped computations. This
was prone to errors (e.g. due to merged or split segments). The data is
now aligned in three columns (source, target, mt) and allows the use of
Pandas convenience functions to avoid code smell. Also, JSON input style
has changed so that PED details are no longer given as a list. The
new dictionary format is easier to create and it avoids confusion.